auto_robo

+++ Project_wsl_2026/Software/system_core_ver1/ir_seeeduinoxiao/lib/movement_average/movement_average.cpp
@@ -5,7 +5,7 @@ void Movement_average::set(int average_count)
     _average_count = constrain(average_count, 1, AVERAGE_COUNT_LIMIT);
 }
 
-void Movement_average::add(float data)
+void Movement_average::add(int data)
 {
     _data[_add_index] = data;
 
@@ -17,9 +17,9 @@ void Movement_average::addNone()
     add(0xFFFF);
 }
 
-float Movement_average::output()
+int Movement_average::output()
 {
-    float sum = 0.0f, available_count = 0;
+    int sum = 0, available_count = 0;
     for (int i = 0; i < _average_count; i++)
     {
         if (_data[i] != 0xFFFF)
@@ -29,7 +29,7 @@ float Movement_average::output()
         }
     }
 
-    float ave = 0.0f;
+    double ave = 0.0;
     if (available_count > 0)
     {
         ave = sum / available_count;
