+++ Project_wsl_2026/Software/system_core_ver1/ir_seeeduinoxiao/lib/movement_average/movement_average.hpp
@@ -17,10 +17,10 @@ private:
 public:
     void set(int average_count);
 
-    void add(float data);
+    void add(int data);
     void addNone();
 
-    float output();
+    int output();
 
     bool cant_compute() const { return _cant_compute; }
 };
